% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{bf.create.ts}
\alias{bf.create.ts}
\title{BFTimeSeries
    Create a timeseries object}
\usage{
bf.create.ts(resp)
}
\arguments{
\item{resp}{response object from server}
}
\description{
BFTimeSeries
    Create a timeseries object
}
