% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comms.R
\name{bf.delete}
\alias{bf.delete}
\title{dev bf.delete}
\usage{
bf.delete(client, path, params)
}
\arguments{
\item{client}{BFCLient object}

\item{path}{"character"}

\item{params}{"list"}
}
\description{
dev bf.delete
}
