% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comms.R
\docType{methods}
\name{bf.get,BFClient,character,list-method}
\alias{bf.get,BFClient,character,list-method}
\title{BFClient}
\usage{
\S4method{bf.get}{BFClient,character,list}(client, path, params)
}
\arguments{
\item{client}{BFClient}

\item{path}{"character"}

\item{params}{"list"}
}
\description{
BFClient
}
