% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comms.R
\docType{methods}
\name{bf.put,BFClient,character,character-method}
\alias{bf.put,BFClient,character,character-method}
\title{BFClient}
\usage{
\S4method{bf.put}{BFClient,character,character}(client, path, params)
}
\arguments{
\item{client}{BFClient}

\item{path}{""}

\item{params}{""}
}
\description{
BFClient
}
