% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R, R/core.R
\docType{class}
\name{BFClient-class}
\alias{BFClient-class}
\alias{initialize,BFClient-method}
\alias{bf.datasets,BFClient-method}
\alias{contents,BFClient,BFDataset-method}
\alias{contents.id,BFClient,character-method}
\title{Class BFClient.}
\usage{
\S4method{initialize}{BFClient}(.Object, ...)

\S4method{bf.datasets}{BFClient}(client)

\S4method{contents}{BFClient,BFDataset}(client, target)

\S4method{contents.id}{BFClient,character}(client, id)
}
\arguments{
\item{client}{BFClient object}

\item{target}{"BFDataset" Id of a Dataset, or Collection}

\item{client}{BFClient client}
}
\value{
BFClient object
}
\description{
This is the main interface class for the Blackfynn R
client. It handles the session and should be passed to
all method calls.
}
\section{Methods (by generic)}{
\itemize{
\item \code{initialize}: Initializes the BFClient object. It opens
a Blackfynn session and stores the API token for the
session.

\item \code{bf.datasets}: Returns a list of Datasets for the organization.
This is mroe inof

\item \code{contents}: Lists the content of a dataset

\item \code{contents.id}: Get the contents of any object on Blackfynn

    @param client BFClient client
    @param id "Character" ID of an object in a dataset
    @export
}}

\section{Slots}{

\describe{
\item{\code{scope}}{The scope in the configuration file}

\item{\code{org}}{The organization for the current session}
}}

