% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R, R/core.R
\docType{class}
\name{BFDataPackage-class}
\alias{BFDataPackage-class}
\alias{show,BFDataPackage-method}
\alias{contents,BFClient,BFDataPackage-method}
\title{Class DataPackage}
\usage{
\S4method{show}{BFDataPackage}(object)

\S4method{contents}{BFClient,BFDataPackage}(client, target)
}
\arguments{
\item{object}{BFClient object}

\item{client}{BFClient client}

\item{target}{"BFDataPackage" Id of a Dataset, or Collection}
}
\description{
This is the DataPackage Class.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Shows a representation of a Blackfynn DataPackage object

\item \code{contents}: Lists the content of a package/collection
}}

\section{Slots}{

\describe{
\item{\code{id}}{The ID of the package}

\item{\code{name}}{The name of the package}

\item{\code{type}}{The type of the package}
}}

