% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R, R/core.R
\docType{class}
\name{BFDataset-class}
\alias{BFDataset-class}
\alias{show,BFDataset-method}
\title{Class Dataset}
\usage{
\S4method{show}{BFDataset}(object)
}
\arguments{
\item{object}{BFClient object}
}
\description{
This is the Dataset class.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Shows a representation of a Blackfynn Dataset object.
}}

\section{Slots}{

\describe{
\item{\code{name}}{The name of the dataset}

\item{\code{description}}{The description of the dataset}

\item{\code{id}}{The dataset ID}

\item{\code{size}}{The size of the dataset in MB.}
}}

