% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R, R/core.R, R/timeseries.R
\docType{class}
\name{BFTimeSeries-class}
\alias{BFTimeSeries-class}
\alias{BFTimeSeries-class}
\alias{show,BFTimeSeries-method}
\alias{show,BFTimeSeries-method}
\title{Class Timeseries}
\usage{
\S4method{show}{BFTimeSeries}(object)

\S4method{show}{BFTimeSeries}(object)
}
\arguments{
\item{object}{BFClient object}

\item{object}{BFClient object}
}
\description{
This is the Timeseries Package Class.

This is the Timeseries Channels Class.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Shows a representation of a Blackfynn Timeseries object

\item \code{show}: Shows a representation of a Blackfynn Timeseries object
}}

\section{Slots}{

\describe{
\item{\code{start}}{Starttime of timeseries}

\item{\code{end}}{Endtime of timeseries}

\item{\code{channels}}{List of channels in timeseries package}

\item{\code{id}}{ID of the channel}

\item{\code{name}}{Name of the channel}

\item{\code{start}}{Starttime of timeseries}

\item{\code{end}}{Endtime of timeseries}
}}

