% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{Blackfynn}
\alias{Blackfynn}
\title{Wrapper function BFClient}
\usage{
Blackfynn(...)
}
\arguments{
\item{scope}{String indicating the API key scope}
}
\value{
A Blackfynn Client object which represents an
active session with the Blackfynn platform.
}
\description{
Wrapper function BFClient
}
