% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\docType{methods}
\name{show,BFClient-method}
\alias{show,BFClient-method}
\title{BFClient
    Shows a representation of a Blackfynn Client object.}
\usage{
\S4method{show}{BFClient}(object)
}
\arguments{
\item{object}{BFClient object}
}
\description{
BFClient
    Shows a representation of a Blackfynn Client object.
}
